+++ hl/CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.10.0] - 2023-08-06
 ### Added
 - Added support for `embedded-hal` version `1.0.0-alpha.11` with the `eh1` feature.
 
 ### Changed
 - Changed the name of the `embedded-hal` feature to `eh0`.
 - Changed `Seek<E>::seek` to `Seek::seek<E>`, moving the seek error generic from the `Seek` trait to the `seek` method.
+- Updated `w5500-ll` dependency from `0.10.0` to `0.11.0`.
 
 ## [0.9.0] - 2022-05-03
 ### Added
@@ -109,7 +110,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0-alpha.1] - 2021-01-31
 - Initial release
 
-[Unreleased]: https://github.com/newAM/w5500-rs/compare/hl%2Fv0.9.0...HEAD
+[Unreleased]: https://github.com/newAM/w5500-rs/compare/hl%2Fv0.10.0...HEAD
+[0.10.0]: https://github.com/newAM/w5500-rs/compare/hl%2Fv0.9.0...hl%2Fv0.10.0
 [0.9.0]: https://github.com/newAM/w5500-rs/compare/hl%2Fv0.8.0...hl%2Fv0.9.0
 [0.8.0]: https://github.com/newAM/w5500-rs/compare/hl%2Fv0.7.1...hl%2Fv0.8.0
 [0.7.1]: https://github.com/newAM/w5500-rs/compare/hl%2Fv0.7.0...hl%2Fv0.7.1